refactor: validation 파일 Money, Lotto, Bonus에 통합
